+++ libs/sdk-backend-spi/src/workspace/users/index.ts
@@ -1,55 +1,14 @@
 // (C) 2019-2022 GoodData Corporation
-import { ObjRef } from "@gooddata/sdk-model";
+/* eslint-disable @typescript-eslint/no-empty-interface, import/namespace */
+import * as m from "@gooddata/sdk-model";
 import { IPagedResource } from "../../common/paging";
 
 /**
- * User
+ * Represents platform user in context of the workspace.
+ * @deprecated Use {@link @gooddata/sdk-model#IWorkspaceUser}
  * @public
  */
-export interface IWorkspaceUser {
-    /**
-     * Stored user reference
-     */
-    ref: ObjRef;
-
-    /**
-     * User uri
-     */
-    uri: string;
-
-    /**
-     * Login - unique user ID for logging into the platform
-     */
-    login: string;
-
-    /**
-     * Contact email of the user
-     */
-    email: string;
-
-    /**
-     * Full name.
-     *
-     * Note: This property has higher priority than firstName / lastName.
-     * Backend implementation MUST fill this property if user names are supported.
-     */
-    fullName?: string;
-
-    /**
-     * First name - when backend implementations supports it.
-     */
-    firstName?: string;
-
-    /**
-     * Last name - when backend implementations supports it.
-     */
-    lastName?: string;
-
-    /**
-     * User status - when backend implementations supports it.
-     */
-    status?: "ENABLED" | "DISABLED";
-}
+export interface IWorkspaceUser extends m.IWorkspaceUser {}
 
 /**
  * Configuration options for querying users
